% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFixedPoints.R
\name{addFixedPoints}
\alias{addFixedPoints}
\title{Add fixed points}
\usage{
addFixedPoints(x, y, types, r = 1)
}
\arguments{
\item{x}{the x coordinates}

\item{y}{the y coordinates}

\item{types}{the types of fixed points}

\item{r}{radius of the plotting character}
}
\description{
Add a fixed point to a plot
}
\details{
This function adds fixed points to a plt
}
