% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animate}
\alias{animate}
\title{Animate}
\usage{
animate(dir, gifName, cleanUp = FALSE)
}
\arguments{
\item{dir}{the directory to save the animation}

\item{gifName}{name for the saved gif}

\item{cleanUp}{whether to remove temporary files}
}
\description{
Animate a series of plots (depricated)
}
\details{
This function is in the process of being remade with `saveVideo`
}
