% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexExpand.R
\name{indexExpand}
\alias{indexExpand}
\title{Expand an index}
\usage{
indexExpand(i, j, imax, jmax)
}
\arguments{
\item{i}{first index}

\item{j}{second index}

\item{imax}{the maximum value i can attain}

\item{jmax}{the maximum value j can attain}
}
\description{
A function for indexing subfigures
}
\details{
Adds the appropriate number of 00 before an integer to properly sort them 
when converted to a string
}
